Add UI

+++ Assets/Scripts/DamageControl/BaseDamagable.cs
@@ -2,36 +2,62 @@ using System.Collections;
 using UnityEngine;
 using UnityEngine.Events;
 
-class BaseDamagable : MonoBehaviour
+class BaseDamagable : MonoBehaviour, IResettable
 {
+    public void Reset() 
+    {
+        _isDead = false;
+        
+        if (HealthValueConnector != null) 
+        {
+            Health = HealthValueConnector.Config.DefaultValue;
+            HealthValueConnector.SetValue(Health);
+        }
+    }
+
     [Header("References")]
     [SerializeField] private GameObject _parent;
 
     [field: Header("Stats")]
     [field: SerializeField] public float Health { get; private set; }
+    [Tooltip("Health value connector (optional)")]
+    [SerializeField] private FloatRangeValue HealthValueConnector;
+    [SerializeField] private bool IsDestroyable = true;
 
     [Header("Events")]
-    [SerializeField] protected UnityEvent OnTakeDamage;
+    [SerializeField] protected FloatUnityEvent OnTakeDamage;
     [SerializeField] protected UnityEvent OnDie;
 
     private float _objectDestroyDelay = 0f; 
+    private bool _isDead = false;
 
     public void SetObjectDestroyDelay(float delay) => _objectDestroyDelay = delay;
+
+    private void Start() 
+    {
+        if (HealthValueConnector != null) 
+        {
+            Health = HealthValueConnector.Config.DefaultValue;
+            HealthValueConnector.SetValue(Health);
+        }
+    }
 
     public virtual void TakeDamage(float damage) 
     { 
         Health -= damage;
 
-        OnTakeDamage?.Invoke();
+        OnTakeDamage?.Invoke(damage);
 
-        if (Health <= 0) Die();
+        if (Health <= 0 && !_isDead) Die();
     }
 
     public virtual void Die() 
     { 
+        _isDead = true;
+
         OnDie?.Invoke();
 
-        StartCoroutine(_destroyAfterSeconds(_objectDestroyDelay));
+        if (IsDestroyable) StartCoroutine(_destroyAfterSeconds(_objectDestroyDelay));
     }
 
     private IEnumerator _destroyAfterSeconds(float seconds)
